requirements.txt
spacy==3.8.0
python-dateutil==2.8.2
rich==13.7.0
streamlit==1.32.0
spacy==3.8.0
pandas==2.1.4
python-magic==0.4.27